perf: split client utils, avoid extra fetch, and reduce bundle size.
