+++ src/services/utils.ts
@@ -1,43 +1,14 @@
-import * as d3 from 'd3-format';
-import { format } from 'date-fns';
 import type { ActivityBase } from './activity';
 import { apiRemoteStorage } from './api';
-import { Equipment } from './equipment';
 import { GearStats } from './gear';
 import { getLogger } from './logger';
 import { getActivitySportType } from './strava-sdk';
-
-export type LocalActivity = {
-  lastUpdated: number;
-  activities: ActivityBase[];
-};
 
 export interface RemoteStorageResponse {
   success: boolean;
   data?: unknown;
   error?: string;
   timestamp: Date;
-}
-
-const locale = d3.formatLocale({
-  decimal: ',',
-  thousands: '.',
-  grouping: [3],
-  currency: ['R$', ''],
-});
-
-function secondsToHms(totalSeconds: number) {
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-
-  return `${String(hours).padStart(2, '0')}:${String(minutes).padStart(
-    2,
-    '0',
-  )}`;
-}
-
-function saveLocalStat(value: LocalActivity) {
-  localStorage.setItem('local-stat', JSON.stringify(value));
 }
 
 async function saveRemote(
@@ -81,54 +52,6 @@ async function saveRemote(
       timestamp: new Date(),
     };
   }
-}
-
-function fallbackCopyTextToClipboard(text: string) {
-  var textArea = document.createElement('textarea');
-  textArea.value = text;
-
-  // Avoid scrolling to bottom
-  textArea.style.top = '0';
-  textArea.style.left = '0';
-  textArea.style.position = 'fixed';
-
-  document.body.appendChild(textArea);
-  textArea.focus();
-  textArea.select();
-
-  document.execCommand('copy');
-
-  document.body.removeChild(textArea);
-}
-
-function copyTextToClipboard(text: string) {
-  if (!navigator.clipboard) {
-    fallbackCopyTextToClipboard(text);
-    return;
-  }
-  navigator.clipboard
-    .writeText(text)
-    .then(() => {
-      return;
-    })
-    .catch((err) => getLogger().error({ err }, 'Error copying text'));
-}
-
-function copyEventDetailsToClipboard(
-  equipment: Equipment,
-  distance: number,
-  movingTime: number,
-) {
-  const formattedDate = format(new Date(), 'dd/MM/yyyy');
-  const formattedDistance = locale.format(',.2f')(distance / 1000);
-  const formattedEquipmentDistance = locale.format(',.2f')(
-    (equipment.distance ?? 0) / 1000,
-  );
-  const formattedTime = secondsToHms(movingTime);
-
-  const text = `. ${formattedDate} - ${formattedDistance} km - ${formattedTime} h [${formattedEquipmentDistance} km]`;
-
-  copyTextToClipboard(text);
 }
 
 function updateActivityInArray(
@@ -224,13 +147,8 @@ function mergeGearStats(
 }
 
 export {
-  copyEventDetailsToClipboard,
-  copyTextToClipboard,
-  locale,
   mergeGearStats,
-  saveLocalStat,
   saveRemote,
-  secondsToHms,
   updateActivityInArray,
   mergeActivities,
 };
