+++ src/utils/format.ts
@@ -1,0 +1,20 @@
+import * as d3 from 'd3-format';
+
+const locale = d3.formatLocale({
+  decimal: ',',
+  thousands: '.',
+  grouping: [3],
+  currency: ['R$', ''],
+});
+
+function secondsToHms(totalSeconds: number) {
+  const hours = Math.floor(totalSeconds / 3600);
+  const minutes = Math.floor((totalSeconds % 3600) / 60);
+
+  return `${String(hours).padStart(2, '0')}:${String(minutes).padStart(
+    2,
+    '0',
+  )}`;
+}
+
+export { locale, secondsToHms };
